+++ tests/unit/Fixture/Container/ContainerWithEntryPoint.php
@@ -8,7 +8,7 @@ class ContainerWithEntryPoint extends AbstractContainer
     /**
      * @var string[]
      */
-    protected $entryPoints = [
+    protected static $entryPoints = [
         \WoohooLabs\Zen\Tests\Unit\Fixture\Container\ContainerWithEntryPoint::class => 'WoohooLabs__Zen__Tests__Unit__Fixture__Container__ContainerWithEntryPoint',
         \Psr\Container\ContainerInterface::class => 'Psr__Container__ContainerInterface',
         \WoohooLabs\Zen\Tests\Unit\Double\StubDefinition::class => 'WoohooLabs__Zen__Tests__Unit__Double__StubDefinition',