+++ tests/unit/Fixture/Container/EmptyContainerWithNamespace.php
@@ -8,7 +8,7 @@ class EmptyContainerWithNamespace extends AbstractContainer
     /**
      * @var string[]
      */
-    protected $entryPoints = [
+    protected static $entryPoints = [
         \WoohooLabs\Zen\Tests\Unit\Fixture\Container\EmptyContainerWithNamespace::class => 'WoohooLabs__Zen__Tests__Unit__Fixture__Container__EmptyContainerWithNamespace',
         \Psr\Container\ContainerInterface::class => 'Psr__Container__ContainerInterface',
     ];