+++ tests/unit/Fixture/Container/EmptyContainerWithoutNamespace.php
@@ -7,7 +7,7 @@ class EmptyContainerWithoutNamespace extends AbstractContainer
     /**
      * @var string[]
      */
-    protected $entryPoints = [
+    protected static $entryPoints = [
         \EmptyContainerWithoutNamespace::class => 'EmptyContainerWithoutNamespace',
         \Psr\Container\ContainerInterface::class => 'Psr__Container__ContainerInterface',
     ];